requirements.txt
amqp==2.2.2
awsebcli==3.12.1
billiard==3.5.0.3
blessed==1.14.2
boto3==1.5.14
botocore==1.8.28
celery==4.1.0
cement==2.8.2
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
defusedxml==0.5.0
Django==2.0.1
django-allauth==0.34.0
django-debug-toolbar==1.9.1
django-storages==1.6.5
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
facepy==1.0.9
future==0.16.0
idna==2.6
jmespath==0.9.3
kombu==4.1.0
oauthlib==2.0.6
parameterized==0.6.1
pathspec==0.5.0
pipreqs==0.4.9
psycopg2==2.7.3.2
PyJWT==1.5.3
python-dateutil==2.6.1
python-dotenv==0.7.1
python-twitter==3.3
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.12
requests==2.9.1
requests-oauthlib==0.8.0
s3transfer==0.1.12
semantic-version==2.5.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
sqlparse==0.2.4
tabulate==0.7.5
termcolor==1.1.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
websocket-client==0.46.0
yarg==0.1.9
